threadpool

+++ PARCM_DISCM/FPSCounter.cpp
@@ -46,5 +46,5 @@ void FPSCounter::draw(sf::RenderWindow* targetWindow)
 void FPSCounter::updateFPS(sf::Time elapsedTime)
 {
 	float fps = 1.0f / elapsedTime.asSeconds();
-	this->statsText->setString("FPS: " + std::to_string((int)fps));
+	this->statsText->setString("FPS: " + std::to_string(fps));
 }
